{"schema":"decodex/commit/1","summary":"Modularize state autonomy and review records","authority":"manual"}
